Wip on tweak.ca, fix an assert fail when trying to use a non-function like a function

+++ cuttlefish/src/mesh.cpp
@@ -13,7 +13,7 @@ namespace mesh {
 
 float consume_float(TokenStream& tokens)
 {
-    return atof(tokens.consume(FLOAT_TOKEN).c_str());
+    return (float) atof(tokens.consume(FLOAT_TOKEN).c_str());
 }
 
 int consume_int(TokenStream& tokens)
